rename fuzz_html to html_report
