+++ post-processing/analyses/optimal_targets.py
@@ -25,7 +25,7 @@ from typing import (
 
 import fuzz_analysis
 import data_loader
-import fuzz_html
+import html_report
 import html_helpers
 import utils
 
@@ -294,7 +294,7 @@ class FuzzOptimalTargetAnalysis(fuzz_analysis.AnalysisInterface):
             "will improve reachability such that it becomes:</p>"
         )
         tables.append(f"myTable{len(tables)}")
-        html_string += fuzz_html.create_top_summary_info(
+        html_string += html_report.create_top_summary_info(
             tables,
             new_profile,
             conclusions,
@@ -309,7 +309,7 @@ class FuzzOptimalTargetAnalysis(fuzz_analysis.AnalysisInterface):
                        "functions in the project will be:</p>"
         table_id = "all_functions_overview_table"
         tables.append(table_id)
-        all_function_table, all_functions_json = fuzz_html.create_all_function_table(
+        all_function_table, all_functions_json = html_report.create_all_function_table(
             tables, new_profile, coverage_url, basefolder, table_id)
         html_string += all_function_table
         html_string += "</div>"  # close report-box
